+++ changelog.md
@@ -1,4 +1,7 @@
 # ErysEdits Changelog
+## v1.1.0
+- Add an additional keybind to save and exit an editor level
+- Bump Node IDs minimum requirement to v1.9.0
 ## v1.0.2
 - Rebuild the mod due to *someone* mangling up `FLAlertLayer` codegen
 - Fix funky text scaling behavior in `FLAlertLayer`s with scrolling enabled by default